+++ asyncapi/_tests/test_api.py
@@ -28,11 +28,11 @@ def test_should_get_api(fake_api):
 
 @pytest.mark.asyncio
 async def test_should_build_api_with_servers_bindings(
-    fake_broadcast_cls, mocker, server_bindings_str,
+    fake_events_handler_cls, mocker, server_bindings_str,
 ):
     asyncapi.build_api('fake', server_bindings=server_bindings_str)
 
-    assert fake_broadcast_cls.call_args_list == [
+    assert fake_events_handler_cls.call_args_list == [
         mocker.call('kafka://fake.fake?option1=0.1&option2=0')
     ]
 
@@ -67,28 +67,28 @@ async def test_should_build_api_with_channels_subscribers_new_channel(
 async def test_should_publish_message(
     spec_dict_publish,
     fake_api,
-    fake_broadcast,
+    fake_events_handler,
     fake_publish_message,
     mocker,
     json_message,
 ):
     await fake_api.publish('fake', fake_publish_message)
 
-    assert fake_broadcast.publish.call_args_list == [
+    assert fake_events_handler.publish.call_args_list == [
         mocker.call(channel='fake', message=json_message.decode())
     ]
 
 
 @pytest.mark.asyncio
 async def test_should_listen_message(
-    fake_api, fake_broadcast, mocker, fake_message
+    fake_api, fake_events_handler, mocker, fake_message
 ):
     fake_operation = asynctest.CoroutineMock()
     fake_api.operations[('fake', 'fake_operation')] = fake_operation
 
     await fake_api.listen('fake')
 
-    assert fake_broadcast.subscribe.call_args_list == [
+    assert fake_events_handler.subscribe.call_args_list == [
         mocker.call(channel='fake')
     ]
     assert fake_operation.call_args_list == [mocker.call(fake_message)]
@@ -116,11 +116,15 @@ async def test_should_not_publish_for_invalid_channel(
 
 @pytest.mark.asyncio
 async def test_should_not_listen_for_invalid_message(
-    fake_api, fake_broadcast, mocker, async_iterator, json_invalid_message,
+    fake_api,
+    fake_events_handler,
+    mocker,
+    async_iterator,
+    json_invalid_message,
 ):
     fake_operation = asynctest.CoroutineMock()
     fake_api.operations[('fake', 'fake_operation')] = fake_operation
-    fake_broadcast.subscribe.return_value = async_iterator(
+    fake_events_handler.subscribe.return_value = async_iterator(
         [mocker.MagicMock(message=json_invalid_message)]
     )