+++ asyncapi/_tests/test_api_auto_spec.py
@@ -16,37 +16,37 @@ def test_should_get_api(fake_api):
 
 @pytest.mark.asyncio
 async def test_should_build_api_with_servers_bindings(
-    fake_broadcast_cls, server_bindings_str, mocker,
+    fake_events_handler_cls, server_bindings_str, mocker,
 ):
     asyncapi.build_api_auto_spec(
         'asyncapi._tests', server_bindings=server_bindings_str
     )
 
-    assert fake_broadcast_cls.call_args_list == [
+    assert fake_events_handler_cls.call_args_list == [
         mocker.call('kafka://fake.fake?option1=0.1&option2=0')
     ]
 
 
 @pytest.mark.asyncio
 async def test_should_publish_message(
-    fake_api, fake_broadcast, fake_message, mocker, json_message
+    fake_api, fake_events_handler, fake_message, mocker, json_message
 ):
     await fake_api.publish('fake', fake_message)
 
-    assert fake_broadcast.publish.call_args_list == [
+    assert fake_events_handler.publish.call_args_list == [
         mocker.call(channel='fake', message=json_message.decode())
     ]
 
 
 @pytest.mark.asyncio
 async def test_should_listen_message(
-    fake_api, fake_broadcast, fake_message, mocker
+    fake_api, fake_events_handler, fake_message, mocker
 ):
     fake_operation = asynctest.CoroutineMock()
     fake_api.operations[('fake', 'fake_operation')] = fake_operation
     await fake_api.listen('fake')
 
-    assert fake_broadcast.subscribe.call_args_list == [
+    assert fake_events_handler.subscribe.call_args_list == [
         mocker.call(channel='fake')
     ]
     assert fake_operation.call_args_list == [mocker.call(fake_message)]