+++ docs/src/gcloud_pubsub/py_spec_user_events.py
@@ -1,0 +1,1 @@
+../py_spec_user_events.py