+++ docs/src/gcloud_pubsub/specification.py
@@ -14,8 +14,8 @@ class UserUpdatePayload:
 
 
 dev_server = asyncapi.Server(
-    url='localhost',
-    protocol=asyncapi.ProtocolType.REDIS,
+    url='asyncapi-local',
+    protocol=asyncapi.ProtocolType.GCLOUD_PUBSUB,
     description='Development Broker Server',
 )
 message = asyncapi.Message(
@@ -26,9 +26,6 @@ message = asyncapi.Message(
 )
 user_update_channel = asyncapi.Channel(
     description='Topic for user updates',
-    subscribe=asyncapi.Operation(
-        operation_id='receive_user_update', message=message,
-    ),
     publish=asyncapi.Operation(message=message),
 )
 
@@ -37,6 +34,6 @@ spec = asyncapi.Specification(
         title='User API', version='1.0.0', description='API to manage users',
     ),
     servers={'development': dev_server},
-    channels={'user/update': user_update_channel},
+    channels={'user-update': user_update_channel},
     components=asyncapi.Components(messages={'UserUpdate': message}),
 )