+++ docs/src/gcloud_pubsub/subscriber.sh
@@ -1,0 +1,4 @@
+PUBSUB_EMULATOR_HOST=localhost:8086 PYTHONPATH=. asyncapi-subscriber \
+    --url http://localhost:5000/asyncapi.json \
+    --api-module user_events \
+    --channels-subscribes 'user-update:user-update-custom-sub=receive_user_update'